backend/requirements.txt
aniso8601==9.0.1
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
clickclick==20.10.2
connexion==2.10.0
Flask==2.0.2
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
flask-serialize==2.0.3
greenlet==1.1.2
idna==3.3
inflection==0.5.1
isodate==0.6.1
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==3.2.0
MarkupSafe==2.0.1
PyMySQL[rsa]==1.0.2
openapi-schema-validator==0.1.6
openapi-spec-validator==0.3.3
pydantic==1.9.0
pyrsistent==0.16.1
pytest==7.0.0
python-dotenv==0.19.2
pytz==2021.3
PyYAML==5.4.1
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.31
swagger-ui-bundle==0.0.9
urllib3==1.26.8
Werkzeug==2.0.2
bcrypt==3.2.0
PyJWT==2.3.0